multi-tenant little program functionality 1. register group

+++ src/test/java/com/hellocrypto/test/dao/CertificateDaoTest.java
@@ -42,7 +42,7 @@ public class CertificateDaoTest {
         mockCertificateDao = mock(CertificateDaoImpl.class);
         
         certificate1 = new Certificate();
-        certificate1.setId(1);
+        certificate1.setId(1L);
         certificate1.setName("mock cer 1");
         String cer1Fingerprint = MD5.md5Base64(ByteUtil.parseByte2HexStr(ByteUtil.int2byte(1)));
         certificate1.setCertificateBinary(cerRawBinary);
@@ -50,7 +50,7 @@ public class CertificateDaoTest {
         certificate1.setTimestamp(new Timestamp(new Date().getTime()));
         
         certificate2 = new Certificate();
-        certificate2.setId(2);
+        certificate2.setId(2L);
         certificate2.setName("mock cer 2");
         String cer2Fingerprint = MD5.md5Base64(ByteUtil.parseByte2HexStr(ByteUtil.int2byte(2)));
         certificate2.setCertificateBinary(cerRawBinary);
@@ -58,7 +58,7 @@ public class CertificateDaoTest {
         certificate2.setTimestamp(new Timestamp(new Date().getTime()));
         
         certificate3 = new Certificate();
-        certificate3.setId(3);
+        certificate3.setId(3L);
         certificate3.setName("mock cer 3");
         String cer3Fingerprint = MD5.md5Base64(ByteUtil.parseByte2HexStr(ByteUtil.int2byte(3)));
         certificate3.setCertificateBinary(cerRawBinary);
@@ -84,7 +84,7 @@ public class CertificateDaoTest {
         for (int i = 0; i < certificates.size(); i++) {
             Certificate cer = certificates.get(i);
             System.out.println("certificate: " + cer.getId() + "-" + cer.getName() + "-" + cer.getPubKeyFingerprint());
-            assertEquals(new Integer(i+1), cer.getId());
+            assertEquals(new Long(i+1), cer.getId());
             assertEquals("mock cer " + (i+1), cer.getName());
             assertEquals(MD5.md5Base64(ByteUtil.parseByte2HexStr(ByteUtil.int2byte(i+1))), 
                     cer.getPubKeyFingerprint());
